OD19DEVAT4-37 OD19DEVAT4-36 OD19DEVAT4-26 1)Change comparator of id to general type ObjectEavIdComparator 2)Add parent class for all models ObjectEav 3)Add AttrtypeProperties with constants which are related with Attrtype table in db 4)Hide secret token value in environment variables 5)Change type of id in models to BigInteger

+++ src/main/java/com/netcracker/testerritto/controllers/TestController.java
@@ -4,6 +4,7 @@ import com.netcracker.testerritto.models.Category;
 import com.netcracker.testerritto.models.GradeCategory;
 import com.netcracker.testerritto.services.CategoryService;
 import com.netcracker.testerritto.services.GradeCategoryService;
+import java.math.BigInteger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,12 +18,12 @@ public class TestController {
   private GradeCategoryService gradeCategoryService;
 
   @GetMapping("/category/{categoryId}")
-  public Category getCategory(@PathVariable("categoryId") int id) {
+  public Category getCategory(@PathVariable("categoryId") BigInteger id) {
     return categoryService.getCategoryById(id);
   }
 
   @DeleteMapping("/category/{categoryId}")
-  public void deleteCategory(@PathVariable("categoryId") int id) {
+  public void deleteCategory(@PathVariable("categoryId") BigInteger id) {
     categoryService.deleteCategoryById(id);
   }
 
@@ -32,17 +33,17 @@ public class TestController {
   }
 
   @PostMapping("/category")
-  public int createCategory(Category newCategory) {
+  public BigInteger createCategory(Category newCategory) {
     return categoryService.createCategory(newCategory);
   }
 
   @GetMapping("/category/grade/{id}")
-  public GradeCategory getGradeCategory(@PathVariable int id) {
+  public GradeCategory getGradeCategory(@PathVariable BigInteger id) {
     return gradeCategoryService.getCategoryById(id);
   }
 
   @PostMapping("/category/grade")
-  public int createGradeCategory(GradeCategory newGradeCategory) {
+  public BigInteger createGradeCategory(GradeCategory newGradeCategory) {
     return gradeCategoryService.createGradeCategory(newGradeCategory);
   }
