+++ src/main/java/com/netcracker/testerritto/services/CategoryService.java
@@ -2,6 +2,7 @@ package com.netcracker.testerritto.services;
 
 import com.netcracker.testerritto.dao.CategoryDAO;
 import com.netcracker.testerritto.models.Category;
+import java.math.BigInteger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,11 +11,11 @@ public class CategoryService {
   @Autowired
   private CategoryDAO categoryDAO;
 
-  public Category getCategoryById(int id) {
+  public Category getCategoryById(BigInteger id) {
     return categoryDAO.getCategoryById(id);
   }
 
-  public void deleteCategoryById(int id) {
+  public void deleteCategoryById(BigInteger id) {
     categoryDAO.deleteCategoryById(id);
   }
 
@@ -22,7 +23,7 @@ public class CategoryService {
     return categoryDAO.updateCategory(updatedCategory);
   }
 
-  public int createCategory(Category newCategory) {
+  public BigInteger createCategory(Category newCategory) {
     return categoryDAO.createCategory(newCategory);
   }
 }
